Upstream update

+++ framework/src/main/groovy/org/moqui/impl/context/reference/UrlResourceReference.groovy
@@ -99,7 +99,7 @@ class UrlResourceReference extends BaseResourceReference {
             File f = getFile()
             List<ResourceReference> children = new LinkedList<ResourceReference>()
             for (File dirFile in f.listFiles()) {
-                children.add(new UrlResourceReference().init(dirFile.absolutePath, ecf))
+                children.add(new UrlResourceReference().init("${getLocation()}/${dirFile.getName()}", ecf))
             }
             return children
         } else {
@@ -122,7 +122,9 @@ class UrlResourceReference extends BaseResourceReference {
         }
     }
 
+    @Override
     boolean supportsLastModified() { isFileProtocol }
+    @Override
     long getLastModified() {
         if (isFileProtocol) {
             return getFile().lastModified()
@@ -131,8 +133,16 @@ class UrlResourceReference extends BaseResourceReference {
         }
     }
 
+    @Override
+    boolean supportsSize() { isFileProtocol }
+    @Override
+    long getSize() { isFileProtocol ? getFile().length() : 0 }
+
+    @Override
     boolean supportsWrite() { isFileProtocol }
+    @Override
     void putText(String text) {
+        if (!isFileProtocol) throw new IllegalArgumentException("Write not supported for resource [${getLocation()}] with protocol [${locationUrl?.protocol}]")
         // first make sure the directory exists that this is in
         if (!getFile().parentFile.exists()) getFile().parentFile.mkdirs()
         // now write the text to the file and close it
@@ -141,7 +151,9 @@ class UrlResourceReference extends BaseResourceReference {
         fw.close()
         this.exists = null
     }
+    @Override
     void putStream(InputStream stream) {
+        if (!isFileProtocol) throw new IllegalArgumentException("Write not supported for resource [${getLocation()}] with protocol [${locationUrl?.protocol}]")
         // first make sure the directory exists that this is in
         if (!getFile().parentFile.exists()) getFile().parentFile.mkdirs()
         OutputStream os = new FileOutputStream(getFile())
@@ -151,6 +163,7 @@ class UrlResourceReference extends BaseResourceReference {
         this.exists = null
     }
 
+    @Override
     void move(String newLocation) {
         if (!newLocation) throw new IllegalArgumentException("No location specified, not moving resource at ${getLocation()}")
         ResourceReference newRr = ecf.resource.getLocationReference(newLocation)
@@ -162,5 +175,27 @@ class UrlResourceReference extends BaseResourceReference {
 
         String path = newRr.getUrl().toExternalForm().substring(5)
         getFile().renameTo(new File(path))
+    }
+
+    @Override
+    ResourceReference makeDirectory(String name) {
+        if (!isFileProtocol) throw new IllegalArgumentException("Write not supported for resource [${getLocation()}] with protocol [${locationUrl?.protocol}]")
+        UrlResourceReference newRef = (UrlResourceReference) new UrlResourceReference().init("${location}/${name}", ecf)
+        newRef.getFile().mkdirs()
+        return newRef
+    }
+    @Override
+    ResourceReference makeFile(String name) {
+        if (!isFileProtocol) throw new IllegalArgumentException("Write not supported for resource [${getLocation()}] with protocol [${locationUrl?.protocol}]")
+        UrlResourceReference newRef = (UrlResourceReference) new UrlResourceReference().init("${location}/${name}", ecf)
+        // first make sure the directory exists that this is in
+        if (!getFile().exists()) getFile().mkdirs()
+        newRef.getFile().createNewFile()
+        return newRef
+    }
+    @Override
+    boolean delete() {
+        if (!isFileProtocol) throw new IllegalArgumentException("Write not supported for resource [${getLocation()}] with protocol [${locationUrl?.protocol}]")
+        return getFile().delete()
     }
 }
